+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Prop `closeOnOutsideClick` on `filter-navigator.v3` to close filter on outside click.
 
 ## [3.79.1] - 2020-10-20
 ### Fixed